requirements.txt
pyyaml
opencv-python
pytorch-msssim
tqdm
imageio
scikit-image
einops
timm
onnxruntime-gpu
pytorch-quantization